+++ src/atom/date.ts
@@ -1,0 +1,9 @@
+import { atom } from "recoil";
+
+export const date = atom({
+	key: "date",
+	default: {
+		year: new Date().getFullYear(),
+		month: new Date().getMonth() + 1,
+	},
+});
