+++ src/atom/index.ts
@@ -1,22 +1,4 @@
-import { atom } from "recoil";
+import { date } from "./date";
+import { user } from "./user";
 
-export const user = atom({
-	key: "user",
-	default: {
-		id: "",
-		firstName: "",
-		lastName: "",
-		email: "",
-		phone: "",
-		token: "",
-		isSignin: false,
-	},
-});
-
-export const date = atom({
-	key: "date",
-	default: {
-		year: new Date().getFullYear(),
-		month: new Date().getMonth() + 1,
-	},
-});
+export { date, user };
