+++ src/atom/user.ts
@@ -1,0 +1,14 @@
+import { atom } from "recoil";
+
+export const user = atom({
+	key: "user",
+	default: {
+		id: "",
+		firstName: "",
+		lastName: "",
+		email: "",
+		phone: "",
+		token: "",
+		isSignin: false,
+	},
+});
