+++ src/screens/index.ts
@@ -1,45 +1,4 @@
-import { AccountScreen } from "./AccountScreen";
-import { AccountUpdateScreen } from "./AccountUpdateScreen";
-import { CreditScreen } from "./CreditScreen";
-import { CreditUpdateScreen } from "./CreditUpdateScreen";
-import { HomeScreen } from "./HomeScreen";
 import { ModalScreen } from "./ModalScreen";
 import { NotFoundScreen } from "./NotFoundScreen";
-import { PasscodeScreen } from "./PasscodeScreen";
-import { PasscodeSettingSelectScreen } from "./PasscodeSettingSelectScreen";
-import { PasscodeUpdateScreen } from "./PasscodeUpdateScreen";
-import { PaymentDetailScreen } from "./PaymentDetailScreen";
-import { PaymentListScreen } from "./PaymentListScreen";
-import { SettingSelectScreen } from "./SettingSelectScreen";
-import { SigninActionScreen } from "./SigninActionScreen";
-import { SigninScreen } from "./SigninScreen";
-import { SignupScreen } from "./SignupScreen";
-import { UserInfoRegisterScreen } from "./UserInfoRegisterScreen";
-import { VerifyScreen } from "./VerifyScreen";
-import { VoiceRecordScreen } from "./VoiceRecordScreen";
-import { VoiceRecordSettingSelectScreen } from "./VoiceRecordSettingSelectScreen";
-import { VoiceRecordUpdateScreen } from "./VoiceRecordUpdateScreen";
 
-export {
-	AccountScreen,
-	AccountUpdateScreen,
-	CreditScreen,
-	CreditUpdateScreen,
-	HomeScreen,
-	ModalScreen,
-	NotFoundScreen,
-	PasscodeScreen,
-	PasscodeSettingSelectScreen,
-	PasscodeUpdateScreen,
-	PaymentDetailScreen,
-	PaymentListScreen,
-	SettingSelectScreen,
-	SigninActionScreen,
-	SigninScreen,
-	SignupScreen,
-	UserInfoRegisterScreen,
-	VerifyScreen,
-	VoiceRecordScreen,
-	VoiceRecordSettingSelectScreen,
-	VoiceRecordUpdateScreen,
-};
+export { ModalScreen, NotFoundScreen };
